Add signal 'style-revalidated' to interface XfdashboardStylable
This signal is emitted when the style information for a XfdashboardStylable derived object was invalidated and the new style information was calculated and applied. This signal is mostly useful for stylable (inheriting XfdashboardStylable interface) non-actor (no sub-class of #XfdashboardActor) objects to also invalidate their style information and to get them recalculated and applied if they depend on the stylable object which emitted the signal. Connected XfdashboardDynamicTableLayout layout manager to this signal.

This commit fixes issue GH #168