Create safe function replacement for clutter_actor_get_allocation_box()
Calling clutter_actor_get_allocation_box() is dangerous with animation because it may modify the internal state of the position and size of an actor used for initial values for position and size at the animation to run. Therefore this safe function exists which returns the last allocation set to this actor.