Restore old behaviour to force a theme path but this time via theme search path instead an environment variable.
The behaviour was to take the provided path and add the theme file without any additional subpath like theme name or theme subpath folder. If this built path provides the theme file use it directly although it might not match the theme name (that is for sure as the theme name will not be added to the path built). This behaviour makes development easier to test themes without changing theme by settings or changing symlinks in any of theme searched paths.