Fix back button not working in (virtual) sub-menu "all applications" at root menu section.
We have to force the signal emission of "filter-changed" signal at application menu model (XfdashboardApplicationsMenuModel) as the problem is that the "all applications" menu can only be entered from root section and it is a virtual menu where the applications menu model is not involved at all. So setting the section back to root section again by calling the model's filtering function xfdashboard_applications_menu_model_filter_by_section() with NULL parameter will not emit this signal as nothing changes.

This fixes issue #21